requirements.txt
streamlit
pdfplumber
pandas
openpyxl
sentence-transformers
faiss-cpu
requests
python-dotenv
